modules/price_intelligence/scraper.py
# Scrape eBay, Mercari, Etsy
# modules/price_intelligence/scraper.py

import requests
from bs4 import BeautifulSoup
import time

class PriceScraper:
    
    @staticmethod
    def scrape_ebay(product_name, category=None):
        """
        Scrape eBay sold listings for similar items
        Returns: List of {title, price, sold_date, condition}
        """
        # Search eBay completed listings
        # Filter by sold items only
        # Extract prices, dates, conditions
        pass
    
    @staticmethod
    def scrape_mercari(product_name):
        """Scrape Mercari for similar items"""
        pass
    
    @staticmethod
    def scrape_etsy(product_name, category):
        """Scrape Etsy for vintage/handmade items"""
        pass