requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.4
numpy==1.26.2

# ML/AI Libraries
scikit-learn==1.3.2
scipy==1.11.4

# NLP & Sentiment Analysis
transformers==4.36.2
torch==2.1.2

# API & Validation
marshmallow==3.20.1
python-dotenv==1.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0

gunicorn==21.2.0
flask-limiter==3.5.0
flask-talisman==1.1.0